Accept relative path for erlangPath
Erlang/OTP may not be installed on the host but prepared in a project
specific directory. This directory could depend on where e.g. the git
repository is cloned.
Accept relative path for erlangPath and resolve it based on the
workspace root directory.

Check if erlangPath is a Windows absolute path and discard it on
non-Windows systems. For example in remote development user settings
has a Windows path but the remote host is Linux.

Signed-off-by: Kornel Horvath <kornel.horvath.p.i@gmail.com>